+++ src/types/color.d.ts
@@ -1,0 +1,6 @@
+export type RgbOptions = {
+  alpha?: boolean;
+  decimal?: boolean;
+}
+
+export type HslOptions = Pick<RgbOptions,'alpha'>;