+++ src/types/number.d.ts
@@ -2,3 +2,7 @@ export type IntOptions = {
   min?: number;
   max?: number;
 }
+
+export type FloatOptions = IntOptions & {
+  precision?: number;
+}